+++ playground/mdx/src/env.d.ts
@@ -1,0 +1,4 @@
+declare module "*.mdx" {
+  import { JSX } from "react";
+  export default () => JSX.Element;
+}