+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 3.2.0
+
 - Support HMR for MDX (fixes #52)
 - Fix: when using plugins, apply SWC before esbuild so that automatic runtime is respected for JSX (fixes #56)
 - Fix: use jsxImportSource in optimizeDeps